Use new codebase feature to split triggers for emulator e2e test  (#4448)
Emulator e2e test will now comprise 3 codebase:

1) Codebase that defines v1 triggers

2) Codebase that defines v2 triggers

3) Codebase that defines triggers that activate event/http triggers

This is strictly not necessary but a good way to exercise multi-codebase support in the emulator.

While implementing multi-codebase support, I found a small race condition in the Pubsub Emulator when registering trigger. Reorganizing where we `await` to squash that bug.
