+++ scripts/triggers-end-to-end-tests/triggers/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "functions",
+  "description": "Cloud Functions for Firebase",
+  "scripts": {},
+  "engines": {
+    "node": "12"
+  },
+  "dependencies": {
+    "@google-cloud/pubsub": "^1.1.5",
+    "firebase-admin": "^9.3.0",
+    "firebase-functions": "^3.20.0",
+    "@firebase/database-compat": "0.1.2"
+  },
+  "devDependencies": {
+    "firebase-functions-test": "^0.2.0"
+  },
+  "private": true
+}
