+++ scripts/triggers-end-to-end-tests/v1/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "functions",
+  "description": "Cloud Functions for Firebase",
+  "scripts": {},
+  "engines": {
+    "node": "12"
+  },
+  "dependencies": {
+    "firebase-admin": "^9.3.0",
+    "firebase-functions": "^3.20.0",
+    "@firebase/database-compat": "0.1.2"
+  },
+  "devDependencies": {
+    "firebase-functions-test": "^0.2.0"
+  },
+  "private": true
+}
