+++ scripts/triggers-end-to-end-tests/v2/package.json
@@ -6,10 +6,8 @@
     "node": "12"
   },
   "dependencies": {
-    "@google-cloud/pubsub": "^1.1.5",
     "firebase-admin": "^9.3.0",
-    "firebase-functions": "^3.16",
-    "@firebase/database-compat": "0.1.2"
+    "firebase-functions": "^3.20.0"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.2.0"
